+++ mlrun/platforms/__init__.py
@@ -13,11 +13,10 @@
 # limitations under the License.
 
 import json
+import warnings
 from pprint import pprint
 from time import sleep
 from typing import Optional
-
-import mlrun.runtimes.mounts as mlrun_mounts
 
 from .iguazio import (
     V3ioStreamClient,
@@ -25,17 +24,46 @@ from .iguazio import (
     is_iguazio_session_cookie,
 )
 
+
+class _DeprecationHelper:
+    """A helper class to deprecate old schemas"""
+
+    def __init__(self, new_target: str, version="1.8.0"):
+        self._new_target = new_target
+        self._version = version
+
+    def __call__(self, *args, **kwargs):
+        self._warn()
+        return self._lazy_load()(*args, **kwargs)
+
+    def __getattr__(self, attr):
+        self._warn()
+        return getattr(self._lazy_load(), attr)
+
+    def _lazy_load(self, *args, **kwargs):
+        import mlrun.runtimes.mounts as mlrun_mounts
+
+        return getattr(mlrun_mounts, self._new_target)
+
+    def _warn(self):
+        warnings.warn(
+            f"mlrun.platforms.{self._new_target} is deprecated since version {self._version}, "
+            f"Use mlrun.runtimes.mounts.{self._new_target} instead.",
+            FutureWarning,
+        )
+
+
 # For backwards compatibility
-VolumeMount = mlrun_mounts.VolumeMount
-auto_mount = mlrun_mounts.auto_mount
-mount_configmap = mlrun_mounts.mount_configmap
-mount_hostpath = mlrun_mounts.mount_hostpath
-mount_pvc = mlrun_mounts.mount_pvc
-mount_s3 = mlrun_mounts.mount_s3
-mount_secret = mlrun_mounts.mount_secret
-mount_v3io = mlrun_mounts.mount_v3io
-set_env_variables = mlrun_mounts.set_env_variables
-v3io_cred = mlrun_mounts.v3io_cred
+VolumeMount = _DeprecationHelper("VolumeMount")
+auto_mount = _DeprecationHelper("auto_mount")
+mount_configmap = _DeprecationHelper("mount_configmap")
+mount_hostpath = _DeprecationHelper("mount_hostpath")
+mount_pvc = _DeprecationHelper("mount_pvc")
+mount_s3 = _DeprecationHelper("mount_s3")
+mount_secret = _DeprecationHelper("mount_secret")
+mount_v3io = _DeprecationHelper("mount_v3io")
+set_env_variables = _DeprecationHelper("set_env_variables")
+v3io_cred = _DeprecationHelper("v3io_cred")
 # eof 'For backwards compatibility'
