--- tests/platforms/test_other.py
@@ -1,147 +1,0 @@
-# Copyright 2023 Iguazio
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-import deepdiff
-
-import mlrun
-import mlrun.errors
-import mlrun.runtimes.mounts
-
-
-def test_mount_configmap():
-    expected_volume = {"configMap": {"name": "my-config-map"}, "name": "my-volume"}
-    expected_volume_mount = {"mountPath": "/myConfMapPath", "name": "my-volume"}
-
-    function = mlrun.new_function(
-        "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
-    )
-    function.apply(
-        mlrun.runtimes.mounts.mount_configmap(
-            configmap_name="my-config-map",
-            mount_path="/myConfMapPath",
-            volume_name="my-volume",
-        )
-    )
-
-    assert (
-        deepdiff.DeepDiff(
-            [expected_volume],
-            function.spec.volumes,
-            ignore_order=True,
-        )
-        == {}
-    )
-    assert (
-        deepdiff.DeepDiff(
-            [expected_volume_mount],
-            function.spec.volume_mounts,
-            ignore_order=True,
-        )
-        == {}
-    )
-
-
-def test_mount_hostpath():
-    expected_volume = {"hostPath": {"path": "/tmp", "type": ""}, "name": "my-volume"}
-    expected_volume_mount = {"mountPath": "/myHostPath", "name": "my-volume"}
-
-    function = mlrun.new_function(
-        "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
-    )
-    function.apply(
-        mlrun.runtimes.mounts.mount_hostpath(
-            host_path="/tmp", mount_path="/myHostPath", volume_name="my-volume"
-        )
-    )
-
-    assert (
-        deepdiff.DeepDiff(
-            [expected_volume],
-            function.spec.volumes,
-            ignore_order=True,
-        )
-        == {}
-    )
-    assert (
-        deepdiff.DeepDiff(
-            [expected_volume_mount],
-            function.spec.volume_mounts,
-            ignore_order=True,
-        )
-        == {}
-    )
-
-
-def test_mount_s3():
-    function = mlrun.new_function(
-        "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
-    )
-    function.apply(
-        mlrun.runtimes.mounts.mount_s3(
-            aws_access_key="xx", aws_secret_key="yy", endpoint_url="a.b"
-        )
-    )
-    env_dict = {var["name"]: var["value"] for var in function.spec.env}
-    assert env_dict == {
-        "S3_ENDPOINT_URL": "a.b",
-        "AWS_ACCESS_KEY_ID": "xx",
-        "AWS_SECRET_ACCESS_KEY": "yy",
-    }
-
-    function = mlrun.new_function(
-        "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
-    )
-    function.apply(mlrun.runtimes.mounts.mount_s3(secret_name="s", endpoint_url="a.b"))
-    env_dict = {
-        var["name"]: var.get("value", var.get("valueFrom")) for var in function.spec.env
-    }
-    assert env_dict == {
-        "S3_ENDPOINT_URL": "a.b",
-        "AWS_ACCESS_KEY_ID": {
-            "secretKeyRef": {"key": "AWS_ACCESS_KEY_ID", "name": "s"}
-        },
-        "AWS_SECRET_ACCESS_KEY": {
-            "secretKeyRef": {"key": "AWS_SECRET_ACCESS_KEY", "name": "s"}
-        },
-    }
-
-
-def test_set_env_variables():
-    env_variables = {
-        "some_env_1": "some-value",
-        "SOMETHING": "ELSE",
-        "and_another": "like_this",
-    }
-
-    function = mlrun.new_function(
-        "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
-    )
-    assert function.spec.env == []
-
-    # Using a dictionary
-    function.apply(mlrun.runtimes.mounts.set_env_variables(env_variables))
-    env_dict = {var["name"]: var.get("value") for var in function.spec.env}
-
-    assert env_dict == env_variables
-
-    function = mlrun.new_function(
-        "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
-    )
-    assert function.spec.env == []
-
-    # And using key=value parameters
-    function.apply(mlrun.runtimes.mounts.set_env_variables(**env_variables))
-    env_dict = {var["name"]: var.get("value") for var in function.spec.env}
-
-    assert env_dict == env_variables
